fix : fix compute for mobile platforms #16
